modified 6-size.c file

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long int b;
 	float f;
 
-	printf("Size of char: %lu .\n", sizeof(c));
-	printf("Size of an int: %lu .\n", sizeof(i));
-	printf("Size of a long int: %lu .\n", sizeof(a));
-	printf("Size of a long long int: %lu .\n", sizeof(b));
-	printf("Size of a float: %lu .\n", sizeof(f));
+	printf("Size of char: %lu .\n", (unsigned long)sizeof(c));
+	printf("Size of an int: %lu .\n", (unsigned long)sizeof(i));
+	printf("Size of a long int: %lu .\n", (unsigned long)sizeof(a));
+	printf("Size of a long long int: %lu .\n", (unsigned long)sizeof(b));
+	printf("Size of a float: %lu .\n", (unsigned long)sizeof(f));
 	return (0);
 }
